Added Singleton to ProfileService.ProfileShared. Moved profile struct to ProfileModels
