+++ imageFeed/imageFeed/SplashViewController/SplashViewController.swift
@@ -7,7 +7,7 @@ final class SplashViewController: UIViewController{
     private let showAuthScreenSegue = "ShowAuthenticationScreen"
     private let outh2Service = OAuth2Service.shared
     private let storage = OAuth2TokenStorage()
-    private let profileService = ProfileService()
+
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -19,7 +19,7 @@ final class SplashViewController: UIViewController{
         super.viewDidAppear(animated)
         
         if let token = OAuth2TokenStorage().token{
-            profileService.fetchProfile(token) { [weak self] result in
+            ProfileService.sharedProfile.fetchProfile(token) { [weak self] result in
                 guard let self = self else {return}
                 
                 switch result {
@@ -90,12 +90,12 @@ extension SplashViewController:AuthViewControllerDelegate{
     func authViewController(_ vc: AuthViewController, didAuthWithCode code: String) {
         dismiss(animated: true){ [weak self] in
             guard let self = self else {return}
-            UIBlockingProgressHUD.show()
             fetchOAuthToken(code)
         }
     }
     
     private func fetchOAuthToken(_ code: String){
+        UIBlockingProgressHUD.show()
         OAuth2Service.shared.fetchOAuthToken(code){ [weak self] result in
             guard let self = self else {return}
             switch result {
@@ -107,7 +107,7 @@ extension SplashViewController:AuthViewControllerDelegate{
         }
     }
     private func fetchProfile(_ token: String){
-        profileService.fetchProfile(token) { [weak self] result in
+        ProfileService.sharedProfile.fetchProfile(token) { [weak self] result in
             guard let self = self else {return}
             switch result {
             case .success(let profile):
